Improve character physics and sprite flipping
You no longer "stick perfectly" when landing on a train car. This caused
inconsistencies between landing on a train (even when stopped) and
landing on the ground, and generally screwed up relativity.
FIXED: when you jump off of a train backwards your sprite flips because
of the velocity added to vx
NOTE: you can no longer jump straight up and down on a car and stay in
the same place relative to it.